Feat: 컴포넌트 제작 및 적용
Button, Inner, Title 컴포넌트 제작 및 적용

+++ src/pages/Todo/TodoListItem.js
@@ -1,10 +1,10 @@
-import React from "react";
 import styled from "styled-components";
 import axios from "axios";
 import { useEffect, useState } from "react";
 
 // 컴포넌트import
-import ButtonComponent from "../../components/button";
+import TodoListUpdate from "./TodoUpdate";
+import TodoList from "./TodoList";
 
 const Cont = styled.ul`
   width: 60%;
@@ -15,25 +15,29 @@ const Cont = styled.ul`
 
 const List = styled.li`
   display: flex;
+  margin: 20px 0 0 0;
   justify-content: space-between;
   align-items: center;
 `;
 
-const TodoItem = styled.p``;
-
 export default function TodoListItem() {
   const access_token = window.localStorage.getItem("token");
   const [item, setItem] = useState();
-  const [id, setId] = useState();
+  const [updataId, setUpdateId] = useState();
+  const [updateVal, setUpdateVal] = useState("");
 
+  // [item] 넣어주면 바로 렌더링은 되는데, 계속 데이터 불러와짐
+  // 일단 window.location.reload(); 넣어서 해결함
   useEffect(() => {
     getTodoItem();
-  }, [item]);
+  }, []);
 
   const headers = {
     Authorization: `Bearer ${access_token}`,
+    "Content-Type": "application/json",
   };
 
+  // get
   const getTodoItem = async () => {
     const url = `https://pre-onboarding-selection-task.shop/todos`;
 
@@ -45,18 +49,46 @@ export default function TodoListItem() {
     }
   };
 
+  // delete
   const deleteTodoList = async (props) => {
-    const id = props;
+    const id = props.id;
     const url = `https://pre-onboarding-selection-task.shop/todos/${id}`;
     try {
-      const result = await axios.delete(url, { headers: headers });
+      if (window.confirm("삭제하시겠습니까?")) {
+        const result = await axios.delete(url, { headers: headers });
+        window.location.reload();
+      }
     } catch (err) {
       console.log(err);
     }
   };
 
-  const updateTodoList = async () => {
-    console.log("수정");
+  const val = {
+    todo: updateVal,
+    isCompleted: true,
+  };
+
+  const submitUpdateVal = async (props) => {
+    const id = props.id;
+    const url = `https://pre-onboarding-selection-task.shop/todos/${id}`;
+    try {
+      if (window.confirm("수정하시겠습니까?")) {
+        window.location.reload();
+        const result = await axios.put(url, val, { headers: headers });
+        setUpdateId(0);
+      }
+    } catch (err) {
+      console.log(err);
+    }
+  };
+
+  const hadleUpdateTodo = async (e) => {
+    setUpdateVal(e.target.value);
+  };
+
+  const startUpdateTodo = async (props) => {
+    setUpdateId(props.id);
+    setUpdateVal(props.todo);
   };
 
   return (
@@ -65,25 +97,31 @@ export default function TodoListItem() {
         item.map((el) => {
           return (
             <List key={el.id}>
-              <TodoItem>{el.todo}</TodoItem>
-              <div>
-                <ButtonComponent btnName="수정" method={updateTodoList}>
-                  수정
-                </ButtonComponent>
-                <ButtonComponent
-                  btnName="삭제"
-                  method={() => {
-                    deleteTodoList(el.id);
-                  }}
-                >
-                  삭제
-                </ButtonComponent>
-              </div>
+              <>
+                {updataId == el.id ? (
+                  <TodoListUpdate
+                    updateVal={updateVal}
+                    hadleUpdateTodo={hadleUpdateTodo}
+                    submitUpdateVal={() => {
+                      submitUpdateVal({ id: el.id });
+                    }}
+                  />
+                ) : (
+                  <TodoList
+                    id={el.id}
+                    todo={el.todo}
+                    startUpdateTodo={() => {
+                      startUpdateTodo({ id: el.id, todo: el.todo });
+                    }}
+                    deleteTodoList={() => {
+                      deleteTodoList({ id: el.id });
+                    }}
+                  />
+                )}
+              </>
             </List>
           );
         })}
     </Cont>
   );
 }
-
-//  method={() => {deleteTodoList()}}
